Added more ground (heigth and width)
It wasn't possible to sculpt ground lower as it was at the lowest point.
Doubled the heigth and moved the terrain so that it stands in the
middle. Effective changes are terrain looks exactly the same but can be
sculpted downwards. Also made terrain 1000x1000 with the middle being
position 0, 0, 0.